requirements.txt
streamlit==1.37.1
newspaper3k==0.2.8
requests==2.31.0
beautifulsoup4==4.12.3
lxml==4.9.3
transformers==4.44.2
torch>=2.0.0
sentencepiece==0.2.0
playwright
ftfy
gTTS